Handle image 'object' that is not a string
The image record's 'object' property can be an array of strings, or
could be missing, or could theoretically be something else we don't
expect, like an object.  Handle all of these cases.

Fixes https://issues.dp.la/issues/8539.  The immediate task with this
ticket is to take the first element of an array, if 'object' is an
array.